pyproject.toml
[project]
name = "bufo-spread-modeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "imagehash>=4.3.2",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
    "requests",
    "tqdm>=4.67.3",
]
